Add {un}listen{-chan}! to connection. Fixes #61.
This introduces an unblocking-buffer? backed channel implementation,
which is different than both DataScript's implementation and Datomic's
implementation.  We expose an API similar to DataScript's API.
Datomic's API exposes a queue, which the Peer is responsible for
draining.  (That is, when used it will grow without bound unless the
Peer drains it.)  The likelihood of error with that approach is low: we
ensure any channel provided is non-blocking, and provide a safe wrapper
for the simple case of a single function callback.